A couple of signed/unsigned fixes.
git-svn-id: file:///home/cminyard/t2/svn/openipmi/trunk@1645 a860ddb0-99b4-4eb6-8282-105f033f2ff5

+++ lib/ipmi_lan.c
@@ -4232,7 +4232,7 @@ session_privilege_set(ipmi_con_t *ipmi, ipmi_msgi_t *rspi)
 	goto out;
     }
 
-    if (lan->cparm.privilege != (msg->data[1] & 0xf)) {
+    if (lan->cparm.privilege != (unsigned int) (msg->data[1] & 0xf)) {
 	/* Requested privilege level did not match. */
         handle_connected(ipmi, EINVAL, addr_num);
 	goto out;
